Checks CMake's Libc versions and installs if necessary
I was having an error with Cmake being built on a computer with
a newer version of libc. When trying to run this script on another
computer with the same cmake program, cmake failed to run because
of the computer did not have a new enough version of libc.

This code should now check for this and download/install it if
necessary.