Adds a script that sets up ssh for GNUroot Debian
GNURoot Debian is an app that allows one to run Debian on an Android
device, by using proot (a chroot clone that doesn't require root
access in the first place).
As it is very minimal (and typing on a terminal on a tiny 3.5-inch
touchscreen is awful), this script creates an ssh server.